feat: manage loading
