+++ packages/app-folders/src/components/Tree/Loader/index.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import { Skeleton } from "@webiny/ui/Skeleton";
+
+import { SkeletonWrapper } from "./styled";
+
+export const Loader = () => {
+    return <Skeleton count={4} inline={true} height={"100%"} wrapper={SkeletonWrapper} />;
+};
