+++ packages/app-folders/src/components/Tree/Loader/styled.tsx
@@ -1,0 +1,11 @@
+import styled from "@emotion/styled";
+
+export const SkeletonWrapper = styled("div")`
+    width: 75%;
+    margin: 0 16px 8px;
+    height: 24px;
+
+    &:last-of-type {
+        margin-bottom: 0;
+    }
+`;
